# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Docmaster API",
        "description": "Docmaster API Documentation",
        "contact": {
            "name": "LJN",
            "email": "jordanndjock07@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/abonnement/{id}/subscribe": {
            "post": {
                "tags": [
                    "Abonnements"
                ],
                "summary": "Souscrire à un abonnement",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "8269136dadec6459e8dcd5683e6a9637",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'abonnement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "montant": {
                                        "type": "integer",
                                        "example": "10000"
                                    },
                                    "tel": {
                                        "type": "string",
                                        "example": "237678788778"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "MTN_MOMO | ORANGE_MONEY"
                                    },
                                    "transactionable_type": {
                                        "type": "string",
                                        "example": "abonnement"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Abonnement souscrit avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TransactionAbonnement"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Abonnement souscrit avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Échec de la souscription"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TransactionAbonnement": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "uuid-transaction-1"
                    },
                    "user_id": {
                        "type": "string",
                        "example": "uuid-user-1"
                    },
                    "transactionable_id": {
                        "type": "string",
                        "example": "uuid-obj"
                    },
                    "transactionable_type": {
                        "type": "string",
                        "example": "App\\Models\\Abonnement"
                    },
                    "statut": {
                        "type": "string",
                        "example": "SUCCESS"
                    },
                    "montant": {
                        "type": "number",
                        "example": 10000
                    },
                    "identifiant": {
                        "type": "string",
                        "example": "TXN-2024-0001"
                    },
                    "supprime": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}